backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.4.0
python-multipart>=0.0.6
python-dotenv==1.0.0
strands-agents[gemini]

